Bug 2065545 - [firefox-devtools-mcp] Move cache behavior into src/firefox/cache.ts
